Fix byte order for Duration of Stay in Visa
While the official VDS specification states:

> Throughout this document, we assume Big Endian encoding, i.e.
> byte sequences are read from left to right.

The sample for the Duration of Stay member for Visas is clearly
Little Endian in the very same document.

So Duration of Stay is now interpreted in Little Endian byte order
as this seems to be the correct way.

Also the corresponding functions for reading integers from VDS
byte arrays are now suffixed with Little/Big Endian to make byte
order more explicit.

Parsing a date from a Big Endian UInt24 has been simplified.